+++ packages/ui/ViewManager/View.js
@@ -122,7 +122,17 @@ class View extends React.Component {
 		 * @type {Boolean}
 		 * @default false
 		 */
-		reverseTransition: PropTypes.bool
+		reverseTransition: PropTypes.bool,
+
+		/**
+		 * When `true`, indicates the current locale uses right-to-eft reading order.
+		 *
+		 * The effect depends on how the provided `arranger` handles this option.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 */
+		rtl: PropTypes.bool
 	}
 
 	static defaultProps = {
@@ -231,7 +241,7 @@ class View extends React.Component {
 	 * @private
 	 */
 	prepareTransition = (arranger, callback, noAnimation) => {
-		const {duration, index, previousIndex = index, reverseTransition} = this.props;
+		const {duration, index, previousIndex = index, reverseTransition, rtl} = this.props;
 
 		// Need to ensure that we have a valid node reference before we animation. Sometimes, React
 		// will replace the node after mount causing a reference cached there to be invalid.
@@ -245,6 +255,7 @@ class View extends React.Component {
 				from: previousIndex,
 				node: this.node,
 				reverse: reverseTransition,
+				rtl,
 				to: index,
 				fill: 'forwards',
 				duration